+++ HISTORY.md
@@ -6,6 +6,8 @@
 
 - NEW: Implement feature request [#165](https://github.com/kataras/iris/issues/165). Routes can now be selected by `a custom name`, and this allows us to use the {{ url "custom-name" "any" "named" "parameters"}}`` inside HTMLEngine's templates. Example [here](https://github.com/iris-contrib/examples/tree/master/templates_9).
 
+- Remove the [x/net/context](https://godoc.org/golang.org/x/net/context), it has been useless after v2.
+
 
 ## 3.0.0-alpha.beta -> 3.0.0-beta